fix(msp): correct UB shift and descriptor re-alloc bug
Two CodeRabbit findings on PR #1159:

1. msp.c: `(1 << desc)` uses a signed literal; shifting into or past
   bit 31 is undefined behaviour in C. Cast to `(uint32_t)1` before
   shifting to match the uint32_t mspArmingDisableFlags target.

2. msp_shared.c: `initSharedMsp()` is called on every new CRSF/MSP
   frame (not just once at boot), so `mspDescriptorAlloc()` was
   invoked repeatedly, leaking a new descriptor slot each frame and
   exhausting the 32-bit bitmask. Guard with a one-time init flag so
   the shared-telemetry descriptor is allocated exactly once.

+++ src/main/interface/msp.c
@@ -156,12 +156,12 @@ static uint32_t mspArmingDisableFlags = 0;
 
 static void mspArmingDisableByDescriptor(mspDescriptor_t desc)
 {
-    mspArmingDisableFlags |= (1 << desc);
+    mspArmingDisableFlags |= ((uint32_t)1 << desc);
 }
 
 static void mspArmingEnableByDescriptor(mspDescriptor_t desc)
 {
-    mspArmingDisableFlags &= ~(1 << desc);
+    mspArmingDisableFlags &= ~((uint32_t)1 << desc);
 }
 
 static bool mspIsMspArmingEnabled(void)
